:sparkles: Enemy Movement
The enemy now has the ability to move around (ON PRESS ENTER). This will be handled by an AI on turn order soon, but is implemented as an option for the enemy right now.

There is a new map! Laying the groundwork for Astar movement and building a grid for the enemy to path on...