:sparkles: :construction: Pathfinder Gives Directions
The Navigation Service will now build a map of valid nodes and provide an AStar calculated path between any two points on the map.

- Added callable get_directions() function to NavigationService
- Added AStar3D service to NaviagationService to calculate positions
- Game_Controller has had the duty of building the nav map revoked, and is not a mediator between the grid map and the nav service to build the map and distribute directions.

The only catch is that previously excluded tiles such as pitfall are now included, the grid map only gives coordinates relative to the internal grid coordinates (I think, it seems like this is the case), and as such does not give directions that can be directly followed in the world map.

Next steps for this feature are to hook up enemies to the directions system and re-include pitfalls correctly.